Some Refactor 😉
