+++ src/Components/FrontBody.js
@@ -1,0 +1,164 @@
+import React from "react";
+
+const FrontBody = () => {
+  return (
+    <svg
+      width="70%"
+      height="100%"
+      viewBox="0 0 672 1200"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <g className="body-map__muscle" id="abdominals">
+        <a href="Abdominals.html">
+          <path
+            d="M384.32 490.61C386.41 457.79 388.65 408.87 388.74 397.67C388.79 392.09 388.67 389.6 388.58 387.74L388.56 387.35C388.5 385.83 388.44 384.62 388.54 382.23C388.56 381.7 388.59 381.18 388.62 380.71C388.64 380.58 388.64 380.45 388.64 380.34C388.68 379.72 388.71 379.12 388.76 378.55L388.81 378.04C388.83 377.82 388.84 377.62 388.88 377.39C389.59 370 391.12 365.02 392.36 361.01L392.47 360.64C393.98 355.77 394.81 353.08 393.26 349.6C391.24 345.04 385.88 341.08 376.39 337.12C376.36 337.11 376.32 337.1 376.29 337.09C376.29 337.09 376.29 337.09 376.26 337.07C374.72 336.53 373.1 335.9 371.46 335.19C371.42 335.17 371.39 335.16 371.33 335.14C360.54 331.16 346.9 328.38 340.97 332.51C338.84 334 337.8 336.3 337.8 339.54C337.8 340.28 337.2 340.88 336.42 340.89C336.33 340.86 336.19 340.85 336.1 340.89C335.36 340.89 334.76 340.29 334.76 339.55C334.76 336.32 333.72 334.02 331.57 332.52C329.69 331.21 321.97 327.54 301.28 335.13L301.18 335.17L301.1 335.2C299.5 335.9 297.87 336.54 296.23 337.1H296.21C296.21 337.1 296.17 337.12 296.16 337.13C286.67 341.09 281.31 345.05 279.31 349.61C277.75 353.12 278.59 355.82 280.11 360.73C281.31 364.54 282.95 369.77 283.7 377.43C283.83 378.3 283.91 379.28 283.93 380.4C283.94 380.48 283.94 380.57 283.94 380.7C283.98 381.21 284.01 381.72 284.03 382.25C284.13 384.82 284.09 386.01 283.99 387.8C283.9 389.62 283.78 392.11 283.83 397.69C283.94 409.23 286.19 458.22 288.25 490.63C291.6 543.3 323.93 583.5 336.14 583.49C336.19 583.51 336.25 583.51 336.31 583.51C336.36 583.51 336.41 583.51 336.45 583.49C348.6 583.49 380.92 543.3 384.28 490.62L384.32 490.61ZM291.68 377.57C291.08 377.57 290.56 377.18 290.39 376.6C290.18 375.9 290.59 375.16 291.3 374.94C305.76 370.74 315.58 368.2 330 368.49C334.47 368.47 334.8 361.56 334.8 351.03C334.8 350.29 335.4 349.69 336.18 349.69C336.27 349.72 336.41 349.73 336.49 349.69C337.23 349.69 337.83 350.29 337.83 351.03C337.83 361.57 338.17 368.47 342.66 368.49C357.04 368.2 366.87 370.74 381.31 374.93C381.66 375.04 381.94 375.27 382.12 375.58C382.29 375.89 382.33 376.25 382.23 376.59C382.04 377.3 381.26 377.71 380.57 377.51C367.11 373.59 357.74 371.15 344.95 371.15C344.21 371.15 343.46 371.15 342.7 371.17C339.84 371.15 337.89 369.72 336.73 366.78C336.6 366.45 336.05 366.45 335.92 366.78C334.74 369.76 332.86 371.16 330 371.17C315.97 370.87 306.3 373.36 292.05 377.51C291.95 377.54 291.83 377.56 291.69 377.56L291.68 377.57ZM369.88 378.6C369.79 378.95 369.56 379.25 369.25 379.43C368.95 379.6 368.59 379.65 368.25 379.55L366.97 379.22C361.69 377.83 352.87 375.5 345.68 376.72C339.05 377.85 337.84 381.39 337.84 387.95C337.84 388.69 337.24 389.29 336.46 389.29C336.41 389.27 336.35 389.27 336.29 389.27C336.24 389.27 336.19 389.27 336.15 389.29C335.4 389.29 334.81 388.7 334.81 387.95C334.81 381.39 333.59 377.86 326.95 376.72C319.75 375.5 310.9 377.84 305.62 379.23L304.37 379.55C304.03 379.64 303.68 379.6 303.37 379.43C303.06 379.25 302.83 378.96 302.74 378.6C302.55 377.87 302.98 377.15 303.71 376.96L304.92 376.64C310.41 375.18 319.59 372.73 327.41 374.07C331.62 374.78 334.41 376.51 335.94 379.35C336.09 379.63 336.56 379.63 336.71 379.35C338.24 376.51 341.03 374.78 345.24 374.07C353.03 372.73 362.22 375.18 367.71 376.64L368.93 376.96C369.65 377.15 370.07 377.87 369.88 378.59V378.6ZM293.5 423.27C293.25 423.01 293.12 422.67 293.12 422.31C293.12 421.93 293.28 421.57 293.55 421.33C293.8 421.1 294.13 420.98 294.47 421L295.96 421.02C310.93 421.27 325.07 421.51 332.23 417.45C334.4 416.23 334.82 413.46 334.95 410.87C334.88 409.36 334.92 407.73 334.97 406.17L334.92 404.53C334.83 402.44 334.79 400.7 334.79 399.06C334.79 398.32 335.39 397.72 336.17 397.72C336.26 397.75 336.39 397.75 336.48 397.72C337.22 397.72 337.82 398.32 337.82 399.06C337.82 401.02 337.74 402.97 337.68 404.58C337.66 405.03 337.65 405.46 337.65 405.91C337.7 407.59 337.74 409.37 337.65 411.07C337.85 414.58 338.67 416.48 340.37 417.44C347.53 421.5 361.67 421.26 376.63 421.01L378.15 420.99C378.48 420.97 378.79 421.08 379.04 421.31C379.32 421.56 379.48 421.93 379.49 422.29C379.49 423.03 378.89 423.65 378.16 423.67L376.69 423.69C373.67 423.75 369.73 423.81 365.99 423.81C355.71 423.81 345.32 423.32 339.04 419.77C338.05 419.2 337.25 418.45 336.65 417.55C336.57 417.43 336.43 417.35 336.29 417.35C336.14 417.35 336 417.43 335.93 417.55C335.36 418.45 334.59 419.18 333.56 419.77C327.29 423.33 316.9 423.82 306.61 423.82C302.86 423.82 298.93 423.75 295.93 423.7L294.44 423.68C294.09 423.68 293.75 423.53 293.5 423.27ZM369.92 427.61C369.87 427.98 369.68 428.31 369.39 428.53C369.12 428.74 368.78 428.82 368.43 428.77L368.25 428.75C362.71 428.11 353.43 427.04 345.29 427.17C338.96 427.24 337.83 430.51 337.83 439.66C337.83 440.4 337.23 441 336.45 441C336.4 440.98 336.34 440.97 336.28 440.97C336.23 440.97 336.18 440.97 336.14 440.99C335.4 440.99 334.8 440.39 334.8 439.65C334.8 430.5 333.67 427.23 327.32 427.16C319.22 427.04 309.97 428.1 304.44 428.73L304.17 428.76C303.44 428.85 302.78 428.33 302.69 427.6C302.61 426.86 303.14 426.19 303.87 426.11C309.51 425.45 318.94 424.36 327.36 424.48C331.61 424.54 334.41 426 335.93 428.93C336.08 429.22 336.55 429.22 336.7 428.93C338.22 426 341.02 424.54 345.27 424.48C353.81 424.37 363.52 425.5 368.75 426.11C369.48 426.19 370.01 426.86 369.93 427.59L369.92 427.61ZM295 470.9C294.76 470.64 294.63 470.3 294.64 469.95C294.65 469.58 294.82 469.23 295.09 468.98C295.34 468.75 295.68 468.64 296.02 468.66C301.12 468.84 304.81 469.09 308.07 469.3L309.04 469.36C314.3 469.75 318.85 470.08 327.57 469.82C333.52 469.39 334.79 467.8 334.79 451.93C334.79 451.19 335.39 450.59 336.17 450.58C336.26 450.61 336.4 450.62 336.49 450.58C337.23 450.58 337.83 451.18 337.83 451.92C337.83 468.54 339.46 469.4 345.14 469.83C353.76 470.07 358.3 469.73 363.56 469.35L364.53 469.28C367.79 469.06 371.48 468.81 376.6 468.63C376.92 468.6 377.24 468.71 377.5 468.94C377.79 469.19 377.96 469.55 377.97 469.92C377.98 470.27 377.85 470.61 377.61 470.87C377.36 471.13 377.03 471.28 376.67 471.3C371.13 471.51 367.25 471.78 363.75 472.03C359.57 472.32 355.97 472.58 350.66 472.58C348.59 472.58 346.79 472.55 344.99 472.49C340.44 472.15 338.04 470.74 336.73 467.64C336.66 467.48 336.5 467.37 336.33 467.37C336.16 467.37 336 467.47 335.93 467.63C334.6 470.73 332.21 472.14 327.7 472.48C325.86 472.54 324.02 472.57 321.93 472.57C316.61 472.57 312.84 472.3 308.86 472.02C305.36 471.77 301.47 471.5 295.94 471.29C295.58 471.28 295.25 471.13 295.01 470.87L295 470.9ZM337.83 570.5C337.83 571.24 337.23 571.84 336.49 571.84H336.14C335.4 571.84 334.8 571.24 334.8 570.5V503.38C334.8 502.63 335.39 502.04 336.14 502.04H336.49C337.23 502.04 337.83 502.64 337.83 503.38V570.5ZM337.66 488.3C337.64 488.38 337.64 488.47 337.66 488.56C337.71 489.72 337.71 490.89 337.71 492.02V492.8C337.71 493.54 337.11 494.14 336.37 494.14H336.25C335.51 494.14 334.91 493.54 334.91 492.8V492.02C334.91 490.87 334.91 489.69 334.96 488.49V488.38C334.78 483.98 334.01 480.91 330.82 480.18C325.59 479 311.19 477.57 304.86 477.57C304.57 477.57 304.3 477.57 304.04 477.57C303.73 477.6 303.41 477.48 303.15 477.25C302.87 476.99 302.7 476.63 302.7 476.26C302.7 475.9 302.83 475.56 303.07 475.3C303.32 475.04 303.65 474.9 304.01 474.89C310.18 474.76 325.81 476.29 331.41 477.56C333.44 478.03 334.93 479.08 335.95 480.77C336.03 480.9 336.17 480.98 336.32 480.98C336.47 480.98 336.62 480.9 336.69 480.77C337.69 479.08 339.16 478.04 341.2 477.56C346.75 476.3 362.39 474.78 368.62 474.89C368.98 474.89 369.31 475.04 369.56 475.3C369.8 475.56 369.93 475.89 369.92 476.25C369.92 476.62 369.75 476.97 369.48 477.22C369.24 477.44 368.94 477.56 368.63 477.56C368.6 477.56 368.58 477.56 368.55 477.56C362.8 477.41 347.33 478.91 341.81 480.16C338.63 480.89 337.84 483.93 337.66 488.29V488.3Z"
+            fill="white"
+          />
+        </a>
+      </g>
+      <g className="body-map__muscle" id="obliques">
+        <a href="Obliques.html">
+          <path
+            d="M278.25 405.1L278.22 405.27L278.23 405.26L278.2 405.42L278.181 405.528C275.415 420.952 272.286 438.408 274.68 454.63C279.61 488.3 267.94 503.64 262.69 508.67C257.5 513.64 249.36 516.54 241.25 516.54C239.37 516.54 237.5 516.38 235.66 516.06C235.43 516.02 235.27 515.81 235.3 515.58C236.6 505.04 237.65 495.4 238.26 488.43C238.68 483.48 239.34 478.52 240.05 473.27L240.08 473.02L240.082 473.006C242.671 453.691 245.348 433.726 235.63 414.4C215.71 376.03 210.63 351.86 209.81 347.41V347.31C209.81 347.26 209.81 347.22 209.83 347.17C210.97 343.85 211.84 340.34 212.44 336.75C213.35 331.07 215.45 317.93 211.12 308.04C211.04 307.85 211.1 307.63 211.27 307.51C211.44 307.39 211.67 307.41 211.82 307.55C214.24 309.84 216.37 312.33 218.14 314.46C219.35 315.91 220.38 317.07 221.37 318.09C221.384 318.104 221.398 318.12 221.413 318.137C221.43 318.157 221.449 318.179 221.47 318.2C221.47 318.21 221.54 318.28 221.54 318.28C231.955 329.911 244.178 340.439 254.971 349.736L255.08 349.83L255.22 349.96L255.238 349.975C268.64 361.529 280.21 371.503 281.06 377.77C281.088 377.911 281.108 378.106 281.127 378.295L281.13 378.33V378.45C281.15 378.5 281.17 378.65 281.17 378.82C281.21 379.09 281.23 379.41 281.25 379.75L281.28 380.33C281.3 380.41 281.3 380.49 281.3 380.56C281.51 386.9 280 395.33 278.25 405.1Z"
+            fill="white"
+          />
+          <path
+            d="M460.77 307.59C460.92 307.45 461.15 307.43 461.32 307.55L461.33 307.56C461.5 307.68 461.56 307.9 461.48 308.09C457.17 317.96 459.26 331.07 460.16 336.67L460.18 336.77C460.78 340.36 461.65 343.87 462.79 347.19C462.81 347.23 462.81 347.28 462.81 347.33V347.43C461.66 353.7 456.4 377.07 436.99 414.44C427.27 433.74 429.95 453.72 432.56 473.03V473.13C433.34 478.83 433.97 483.58 434.39 488.44C434.96 495.23 436.01 504.88 437.33 515.59C437.36 515.82 437.2 516.03 436.97 516.07C435.14 516.39 433.26 516.54 431.39 516.54C423.28 516.54 415.13 513.64 409.94 508.67C404.69 503.63 393.02 488.29 397.95 454.63C400.36 438.26 397.2 420.73 394.4 405.27L394.37 405.08C392.62 395.32 391.11 386.89 391.32 380.55C391.32 380.51 391.32 380.44 391.34 380.36C391.34 380.29 391.35 380.2 391.37 380.11C391.41 379.35 391.48 378.57 391.55 377.8C391.55 377.78 391.55 377.72 391.58 377.63C392.6 371.342 404.153 361.386 417.528 349.861L417.676 349.733C428.442 340.462 440.614 329.981 451.07 318.29L451.14 318.21C451.17 318.17 451.21 318.13 451.24 318.1C452.33 316.97 453.41 315.75 454.46 314.47C456.35 312.19 458.42 309.78 460.77 307.59Z"
+            fill="white"
+          />
+        </a>
+      </g>
+      <g className="body-map__muscle" id="forearms">
+        <a href="Forearms.html">
+          <path
+            d="M135.05 418.55C141.32 418.55 148.19 416.1 154.92 411.47V411.46C155.1 411.34 155.33 411.37 155.48 411.52C155.63 411.67 155.65 411.91 155.52 412.08C151.51 417.52 146.59 424.21 144.16 427.73C139.457 435.289 136.843 440.787 134.543 445.625L134.46 445.8L134.457 445.806C130.428 454.264 126.627 462.244 113.47 478.7C101.33 493.83 93.04 500.16 84.26 506.86L84.199 506.906C77.9179 511.692 71.4135 516.648 63.23 525.21C63.21 525.23 63.19 525.25 63.16 525.27C61.56 527.38 58.6 531.44 56.45 535.48C56.4 535.57 56.33 535.64 56.23 535.68C53.2 536.91 50.19 537.52 47.41 537.52C44.92 537.52 42.61 537.03 40.63 536.04C35.84 533.67 32.13 527.92 30.95 521.04C30.94 520.95 30.95 520.86 30.99 520.78C31.81 519.04 32.73 517.24 33.72 515.43C37.33 507.63 41.56 497.2 46.32 485.38L46.43 485.1C57.43 457.87 71.12 423.99 84.49 407.9C92.4 398.38 98.13 392.38 102.32 388L102.5 387.81C108.4 381.64 111.64 378.25 114.13 372.23C114.229 371.995 114.327 371.761 114.425 371.528C114.593 371.127 114.759 370.728 114.93 370.33L115.04 370.08C115.13 369.87 115.36 369.77 115.58 369.84C115.8 369.91 115.92 370.14 115.87 370.36C108.45 400.78 117.14 412.23 125.73 416.48C128.51 417.85 131.64 418.55 135.05 418.55Z"
+            fill="white"
+          />
+          <path
+            d="M638.96 515.58C640.01 517.47 640.88 519.19 641.61 520.78C641.64 520.86 641.65 520.94 641.64 521.03C640.5 527.9 636.8 533.65 631.99 536.05C630.02 537.04 627.7 537.53 625.21 537.53C622.42 537.53 619.4 536.92 616.38 535.7C616.29 535.66 616.21 535.59 616.16 535.5C614.01 531.46 611.05 527.39 609.47 525.33C609.46 525.32 609.443 525.298 609.425 525.275C609.407 525.252 609.39 525.23 609.38 525.22C601.16 516.63 594.65 511.67 588.351 506.871L588.291 506.825C579.533 500.132 571.243 493.796 559.15 478.7C545.92 462.19 542.15 454.23 538.16 445.8L538.02 445.5L538.012 445.484C535.585 440.399 533.075 435.142 528.39 427.63C526.32 424.65 522.31 419.13 517.12 412.09C517 411.92 517.01 411.68 517.16 411.53C517.31 411.37 517.55 411.35 517.72 411.47C524.43 416.1 531.29 418.55 537.57 418.55C540.97 418.55 544.11 417.85 546.89 416.48C555.48 412.24 564.16 400.82 556.8 370.45C556.74 370.23 556.87 370 557.09 369.93C557.31 369.85 557.55 369.97 557.63 370.18L557.633 370.188C557.792 370.575 557.951 370.963 558.12 371.35C558.25 371.64 558.37 371.93 558.49 372.22C561.01 378.3 564.24 381.68 570.12 387.82L570.48 388.2L570.625 388.352C574.773 392.699 580.44 398.639 588.14 407.91C601.55 424.05 615.27 458.06 626.29 485.38L626.52 485.95C631.01 497.07 635.25 507.57 638.96 515.58Z"
+            fill="white"
+          />
+        </a>
+      </g>
+      <g className="body-map__muscle" id="biceps">
+        <a href="Biceps.html">
+          <path
+            d="M207.71 345.02C207.68 345.12 207.61 345.21 207.52 345.26V345.25C207.32 345.37 207.15 345.57 207.04 345.84C202.959 355.849 196 362.113 191.389 366.264L191.36 366.29C190.73 366.85 190.16 367.36 189.67 367.84C187.7 369.66 179.78 379.71 167.95 395.4C155.59 410.5 143.93 415.86 135.1 415.86C131.98 415.86 129.22 415.19 126.91 414.05C117.4 409.37 105.33 392.91 131.34 335.25C131.34 335.24 131.42 335.08 131.42 335.08C140.01 318.85 144.88 313.74 147.79 310.69L148.08 310.39C148.53 309.92 148.94 309.48 149.32 309.03C149.35 309 149.38 308.97 149.41 308.95C152.39 306.96 155.59 305.18 158.94 303.68C161.66 302.46 164.46 301.43 167.25 300.62C167.39 300.59 167.47 300.57 167.53 300.54C170.77 299.61 174.06 298.94 177.38 298.53C177.54 298.53 177.65 298.52 177.76 298.49C179.544 298.291 180.92 298.24 181.741 298.21L181.75 298.21C190.1 297.96 197.98 297.73 203.74 302.43C213.92 310.71 210.92 329.3 209.79 336.29L209.36 336.24L209.79 336.31C209.32 339.25 208.62 342.18 207.71 345.02Z"
+            fill="white"
+          />
+          <path
+            d="M541.2 335.1C567.3 392.89 555.23 409.38 545.71 414.07L545.73 414.08C543.41 415.23 540.64 415.9 537.52 415.9C528.73 415.9 517.12 410.57 504.8 395.57C504.77 395.54 504.75 395.51 504.74 395.48C492.87 379.77 484.95 369.71 482.96 367.86C482.46 367.4 481.9 366.89 481.28 366.34C476.66 362.2 469.7 355.93 465.59 345.87C465.48 345.61 465.3 345.39 465.08 345.25C464.99 345.2 464.93 345.12 464.9 345.02C463.99 342.17 463.29 339.25 462.83 336.34L462.828 336.326C461.696 329.296 458.707 310.725 468.88 302.45C474.656 297.743 482.529 297.98 490.852 298.23L490.87 298.23C491.7 298.26 493.08 298.32 494.81 298.5C494.931 298.524 495.029 298.53 495.117 298.535C495.175 298.539 495.228 298.542 495.28 298.55C498.49 298.94 501.77 299.61 505.04 300.53C505.19 300.59 505.25 300.61 505.32 300.62C508.17 301.44 510.97 302.47 513.69 303.69C517.04 305.19 520.25 306.97 523.22 308.96L523.235 308.971C523.268 308.995 523.293 309.014 523.31 309.04C523.61 309.39 523.92 309.73 524.25 310.08C524.44 310.28 524.639 310.499 524.839 310.719L524.84 310.72L524.862 310.743C527.755 313.766 532.612 318.841 541.2 335.1Z"
+            fill="white"
+          />
+        </a>
+      </g>
+      <g className="body-map__muscle" id="shoulders">
+        <a href="Shoulder.html">
+          <path
+            d="M152.89 303.71C152.83 303.75 152.75 303.77 152.68 303.77H152.66C152.56 303.77 152.47 303.74 152.39 303.68C152.26 303.57 152.19 303.39 152.24 303.22C157.74 282.4 160.82 274.86 162.86 269.87L162.897 269.778C163.919 267.242 164.719 265.259 165.46 262.65C171.07 244.34 184.75 221.38 218.1 216C218.85 215.87 219.39 215.77 219.69 215.7L220.14 215.58C220.467 215.498 220.837 215.404 221.242 215.301C221.728 215.178 222.262 215.042 222.83 214.9C223.469 214.73 224.088 214.571 224.706 214.411L224.71 214.41L225.77 214.14C225.81 214.13 225.84 214.13 225.88 214.13C243.5 210.59 262.01 215.16 271.47 218.2C271.64 218.25 271.77 218.42 271.77 218.6C271.78 218.79 271.66 218.95 271.49 219.02C267.45 220.55 261.53 223.08 256.28 226.42C245.175 233.477 240.444 239.724 233.899 248.368L233.89 248.38C230.84 252.44 227.43 256.97 222.68 262.54L222.658 262.566C217.699 268.376 207.4 280.443 194.34 289.32C190.84 291.69 184.95 293.94 177.31 295.83H177.29C177.24 295.85 177.18 295.86 177.14 295.86C173.74 296.25 170.23 296.96 166.75 297.97C166.733 297.974 166.717 297.977 166.703 297.981C166.604 298.003 166.551 298.015 166.49 298.05C163.47 298.93 160.58 299.99 157.83 301.22C156.1 302.01 154.44 302.84 152.89 303.71Z"
+            fill="white"
+          />
+          <path
+            d="M400.85 218.61C400.85 218.43 400.97 218.27 401.15 218.21L401.14 218.23C410.58 215.19 429.07 210.61 446.76 214.17C447.006 214.234 447.249 214.295 447.49 214.357C447.858 214.451 448.222 214.544 448.59 214.64L449.77 214.94L450.317 215.078C451.36 215.34 452.27 215.569 452.91 215.74C453.2 215.8 453.74 215.91 454.49 216.03C487.82 221.4 501.51 244.33 507.13 262.62C507.872 265.301 508.673 267.288 509.678 269.78L509.71 269.86C511.77 274.9 514.86 282.46 520.36 303.24C520.41 303.41 520.35 303.59 520.21 303.7C520.13 303.76 520.03 303.79 519.94 303.79C519.87 303.79 519.8 303.77 519.73 303.73C518.16 302.85 516.5 302.01 514.78 301.24C507.02 297.76 499.95 296.39 495.38 295.86C487.71 293.99 481.81 291.74 478.26 289.32C465.161 280.424 454.873 268.349 449.914 262.528L449.89 262.5C445.03 256.78 441.4 251.97 438.76 248.45L438.65 248.31C432.13 239.69 427.42 233.47 416.34 226.43C411.09 223.11 405.17 220.57 401.13 219.03C400.96 218.96 400.85 218.79 400.85 218.61Z"
+            fill="white"
+          />
+        </a>
+      </g>
+      <g className="body-map__muscle" id="traps">
+        <a href="Traps.html">
+          <path
+            d="M395.89 217.17C395.84 217.17 395.75 217.15 395.75 217.15V217.14C384.68 213.38 379.06 209.6 379.06 199.34V178.65C379.06 178.48 379.16 178.32 379.32 178.25C379.48 178.18 379.66 178.21 379.79 178.33C382.83 181.15 388.48 186.15 392.19 188.01C395.081 189.475 402.834 192.729 409.776 195.641L410.13 195.79L412.34 196.72C412.95 196.977 413.555 197.233 414.15 197.484C417.946 199.085 421.335 200.515 422.96 201.25C425.92 202.61 430.04 205.78 433.67 208.64L433.91 208.83C434.02 208.918 434.128 209.002 434.233 209.084C434.367 209.188 434.497 209.289 434.62 209.39C434.75 209.51 434.81 209.71 434.75 209.88C434.69 210.06 434.52 210.17 434.34 210.17C433.89 210.16 433.451 210.16 433.001 210.16C415.031 210.16 399.12 215.95 396.05 217.14C396 217.17 395.94 217.17 395.89 217.17Z"
+            fill="white"
+          />
+          <path
+            d="M276.58 217.15C273.44 215.93 256.81 209.87 238.27 210.18V210.19C238.07 210.19 237.91 210.08 237.85 209.9C237.79 209.72 237.85 209.53 237.99 209.41C238.029 209.378 238.07 209.345 238.111 209.312C238.287 209.169 238.477 209.014 238.68 208.86L238.93 208.66L238.933 208.658C241.963 206.278 246.531 202.69 249.64 201.27C251.51 200.43 255.73 198.65 260.2 196.77L262.48 195.81L262.846 195.657C269.784 192.745 277.531 189.494 280.44 188.03C283.94 186.28 289.1 181.8 292.82 178.35C292.95 178.23 293.13 178.2 293.29 178.27C293.45 178.34 293.55 178.5 293.55 178.67V199.36C293.55 209.61 287.93 213.39 276.88 217.16C276.84 217.18 276.79 217.18 276.74 217.18C276.68 217.18 276.63 217.17 276.58 217.15Z"
+            fill="white"
+          />
+        </a>
+      </g>
+      <g className="body-map__muscle" id="chest">
+        <a href="Chest.html">
+          <path
+            d="M277.532 219.782C277.578 219.768 277.624 219.754 277.67 219.74H277.66L277.85 219.68C282.69 218.25 291.6 215.63 307.32 217.12C334.48 219.71 334.79 236.91 334.79 237.64V287.27C334.79 297.21 331.34 306.75 325.08 314.13C318.19 322.23 309.8 328.48 300.15 332.7C300.141 332.703 300.129 332.708 300.115 332.713C300.082 332.727 300.036 332.746 299.98 332.76C298.18 333.44 296.71 334.01 295.37 334.57C295.35 334.58 295.32 334.6 295.32 334.6H295.29C288.29 337.01 281.02 338.18 273.77 338.18C255.58 338.18 237.53 330.82 224.21 317.22C224.14 317.16 224.08 317.1 224.03 317.03L223.98 316.98C223.932 316.928 223.885 316.876 223.838 316.825C223.679 316.653 223.527 316.488 223.38 316.31C223.32 316.23 223.26 316.18 223.21 316.14C223.16 316.11 223.12 316.08 223.09 316.05C222.13 315.02 221.16 313.92 220.2 312.75C213.51 304.62 205.11 295.9 190.2 295.44C190 295.43 189.83 295.28 189.79 295.08C189.75 294.88 189.86 294.68 190.05 294.6C192.28 293.66 194.23 292.63 195.85 291.53C209.29 282.42 219.75 270.14 224.78 264.23C229.196 259.042 232.486 254.704 235.386 250.879L235.59 250.61L236 250.06L236.02 250.034C242.629 241.268 246.991 235.483 257.72 228.68C264.15 224.59 271.91 221.64 275.96 220.25C276 220.23 276.05 220.22 276.11 220.2L276.69 220.02C276.75 219.99 276.79 219.97 276.85 219.97C276.869 219.964 276.887 219.958 276.906 219.954C276.917 219.951 276.929 219.95 276.94 219.95C276.953 219.947 276.967 219.941 276.982 219.936C277.013 219.924 277.049 219.91 277.09 219.91C277.236 219.872 277.382 219.827 277.532 219.782Z"
+            fill="white"
+          />
+          <path
+            d="M372.63 332.76C372.54 332.74 372.44 332.69 372.44 332.69H372.42C362.8 328.48 354.41 322.23 347.51 314.13C341.26 306.77 337.82 297.23 337.82 287.27V237.61C337.81 237.61 337.77 232.46 342.02 227.28C346.71 221.57 354.54 218.15 365.29 217.13C380.986 215.63 389.895 218.259 394.676 219.669L394.68 219.67L394.95 219.75C395.15 219.82 395.35 219.88 395.54 219.93C395.58 219.93 395.63 219.95 395.69 219.97C395.7 219.97 395.79 220 395.79 220C395.83 220.01 395.87 220.02 395.92 220.04C395.94 220.05 396.03 220.08 396.05 220.08C396.16 220.12 396.26 220.15 396.37 220.18L396.51 220.23C396.547 220.241 396.582 220.253 396.616 220.265C396.677 220.287 396.737 220.307 396.8 220.32C400.85 221.72 408.53 224.66 414.88 228.7C425.617 235.498 429.988 241.297 436.595 250.063L436.6 250.07L437.03 250.64C439.98 254.53 443.32 258.93 447.88 264.29C452.95 270.24 463.4 282.49 476.75 291.54C478.37 292.64 480.32 293.67 482.55 294.61C482.74 294.69 482.85 294.89 482.81 295.09C482.78 295.3 482.6 295.45 482.4 295.45C467.48 295.9 459.09 304.62 452.39 312.76C451.52 313.84 450.6 314.86 449.5 316.06C449.5 316.07 449.48 316.09 449.48 316.09C449.47 316.1 449.455 316.112 449.44 316.125C449.425 316.137 449.41 316.15 449.4 316.16C449.34 316.2 449.3 316.23 449.25 316.3L449.03 316.55C448.98 316.603 448.931 316.657 448.882 316.71C448.784 316.817 448.687 316.923 448.58 317.03C448.59 317.04 448.51 317.12 448.44 317.19C435.13 330.8 417.05 338.17 398.85 338.17C391.6 338.17 384.32 337 377.32 334.58H377.28C377.276 334.576 377.271 334.574 377.266 334.572C377.257 334.57 377.246 334.57 377.24 334.57C375.57 333.88 373.9 333.24 372.63 332.76Z"
+            fill="white"
+          />
+        </a>
+      </g>
+      <g className="body-map__muscle" id="quads">
+        <a href="Quads.html">
+          <path
+            d="M458.96 792.3L458.82 793.02L458.79 793.01C455.7 808.78 452.78 823.66 445.26 827.39C442.18 828.92 438.37 828.54 433.62 826.22C431.52 825.19 429.7 825.03 428.21 825.74C425.2 827.18 424.21 831.51 422.96 837L422.956 837.018C420.807 846.414 418.372 857.06 405.71 857.28H405.44C382.64 857.28 375.34 806.56 375.27 806.05C375.26 805.98 375.25 805.92 375.23 805.87C375.2 805.81 375.2 805.76 375.2 805.7C374.55 800.89 373.79 796.12 372.95 791.53V791.38L372.91 791.22C372.88 791.05 372.85 790.88 372.8 790.71C372.43 788.68 372.02 786.51 371.48 783.85L371.42 783.55C371.35 783.23 371.28 782.88 371.22 782.54C365.63 751.25 368.5 747.76 374.79 740.12L374.795 740.114C381.265 732.274 391.032 720.439 399.45 674.95C401.737 662.572 403.351 647.89 405.06 632.351L405.09 632.08L405.118 631.827C409.996 587.785 415.522 537.892 437.54 522.52C437.67 522.43 437.83 522.42 437.97 522.48C438.11 522.54 438.21 522.67 438.23 522.82C441.22 545.99 444.31 565.51 446.09 572.56L446.19 572.92C446.314 573.364 446.463 573.935 446.634 574.597L446.7 574.85C455.6 608.45 474.81 695.9 461.05 782.05V782.07C460.31 785.48 459.62 788.95 458.96 792.3Z"
+            fill="white"
+          />
+          <path
+            d="M267.5 632.08C269.12 646.92 270.81 662.26 273.14 674.95H273.13C281.55 720.45 291.32 732.28 297.79 740.11L297.797 740.118C304.082 747.742 306.958 751.231 301.36 782.54C301.31 782.87 301.24 783.2 301.17 783.53C301.16 783.64 301.14 783.72 301.12 783.79C300.72 785.77 300.33 787.75 299.99 789.7L299.83 790.45C299.77 790.74 299.71 791.04 299.65 791.34C299.64 791.42 299.62 791.52 299.62 791.61C298.77 796.32 298.02 801.04 297.38 805.7C297.38 805.737 297.376 805.774 297.368 805.809C297.364 805.831 297.358 805.851 297.35 805.87C297.343 805.9 297.335 805.94 297.324 805.996C297.32 806.015 297.315 806.037 297.31 806.06C297.3 806.18 295.46 819.01 290.79 831.76C284.68 848.46 276.5 857.28 267.14 857.28H266.89C254.226 857.06 251.792 846.43 249.643 837.044L249.64 837.03L249.6 836.85C248.35 831.44 247.36 827.16 244.37 825.74C242.85 825.03 241.03 825.19 238.96 826.22C234.2 828.54 230.39 828.92 227.31 827.39C219.791 823.661 216.871 808.785 213.782 793.048L213.78 793.04L213.71 792.69L213.601 792.139C212.925 788.731 212.229 785.223 211.48 781.76C197.85 695.79 217.02 608.42 225.89 574.84C226.06 574.22 226.2 573.68 226.32 573.23L226.5 572.55C228.28 565.52 231.37 546.01 234.36 522.84C234.38 522.69 234.48 522.56 234.62 522.5C234.76 522.44 234.93 522.45 235.05 522.54C257.069 537.892 262.594 587.78 267.473 631.824L267.48 631.89L267.5 632.08Z"
+            fill="white"
+          />
+        </a>
+      </g>
+      <g className="body-map__muscle" id="calves">
+        <a href="Calves.html">
+          <path
+            d="M213.13 880.79C213.65 879.35 214.06 878.18 214.36 877.22C219.58 878.81 222.24 881.96 226.36 889.29C225.671 901.305 223.421 919.666 221.458 935.675C220.998 939.426 220.554 943.049 220.15 946.43C218.66 958.9 213.28 974.49 201.61 983.81C201.488 983.915 201.362 984.013 201.237 984.104C201.147 984.169 201.058 984.231 200.97 984.29C196.03 928.87 208.12 894.82 213.13 880.79Z"
+            fill="white"
+          />
+          <path
+            d="M446.26 889.29C450.38 881.96 453.04 878.82 458.26 877.22C458.57 878.18 458.97 879.35 459.49 880.79C464.49 894.82 476.59 928.87 471.65 984.29C471.44 984.15 471.22 983.99 471.01 983.81C459.34 974.48 453.96 958.9 452.47 946.43C452.066 943.049 451.622 939.427 451.162 935.676C449.199 919.667 446.949 901.305 446.26 889.29Z"
+            fill="white"
+          />
+          <path
+            d="M419.322 902.906C418.618 902.747 417.894 902.558 417.19 902.32V902.3C412.25 900.74 407.39 897.7 402.91 893.4C403.53 898.32 402.73 906.81 400.35 927.92C394.241 972.272 398.539 978.278 402.688 984.077L402.69 984.08L402.759 984.178C404.278 986.317 405.847 988.527 406.95 992.28C406.99 992.4 407.02 992.53 407.07 992.65L407.819 992.047C418.013 983.845 426.766 976.803 425.38 954.81C424.524 941.224 422.33 925.11 420.55 912.034L420.546 912.01C420.104 908.758 419.687 905.695 419.322 902.906Z"
+            fill="white"
+          />
+          <path
+            d="M269.71 893.41C265.24 897.71 260.37 900.74 255.43 902.31V902.33C254.72 902.56 253.99 902.76 253.28 902.92C252.915 905.717 252.496 908.791 252.052 912.054L252.05 912.065C250.27 925.14 248.076 941.255 247.22 954.84C245.834 976.833 254.587 983.875 264.781 992.077C265.03 992.277 265.28 992.478 265.53 992.68C265.54 992.656 265.55 992.631 265.559 992.606C265.573 992.569 265.586 992.531 265.598 992.494C265.618 992.432 265.635 992.37 265.65 992.31C266.758 988.53 268.337 986.317 269.862 984.178L269.91 984.11C274.07 978.3 278.36 972.3 272.25 927.95C269.87 906.84 269.07 898.35 269.69 893.43L269.71 893.41Z"
+            fill="white"
+          />
+        </a>
+      </g>
+      <g className="body-map__model">
+        <path
+          d="M342.66 368.05C357.13 367.77 366.96 370.31 381.44 374.53H381.43C382.37 374.8 382.91 375.79 382.64 376.73C382.42 377.51 381.71 378.01 380.94 378.01C380.77 378.01 380.6 377.99 380.44 377.94C366.28 373.82 356.65 371.33 342.68 371.61C339.38 371.59 337.44 369.79 336.31 366.96C335.18 369.79 333.26 371.6 329.97 371.61C315.96 371.33 306.33 373.82 292.17 377.94C292 377.99 291.83 378.01 291.67 378.01C290.9 378.01 290.2 377.51 289.97 376.73C289.7 375.79 290.24 374.8 291.18 374.53C305.66 370.31 315.49 367.77 330 368.05C334.04 368.04 334.35 361.3 334.35 351.03C334.35 350.05 335.15 349.25 336.13 349.25C336.151 349.25 336.172 349.252 336.193 349.256L336.219 349.26L336.224 349.261C336.236 349.263 336.248 349.265 336.261 349.266C336.277 349.268 336.294 349.27 336.31 349.27C336.37 349.26 336.43 349.25 336.49 349.25C337.48 349.25 338.27 350.05 338.27 351.03C338.27 361.3 338.57 368.04 342.66 368.05Z"
+          fill="black"
+        />
+        <path
+          d="M305.74 379.65L304.48 379.98H304.51C303.57 380.23 302.59 379.66 302.35 378.71C302.1 377.77 302.67 376.8 303.62 376.55L304.86 376.22C310.39 374.76 319.64 372.32 327.5 373.66C332.33 374.48 334.94 376.55 336.34 379.17C337.74 376.55 340.35 374.48 345.18 373.66C353.049 372.32 362.288 374.759 367.818 376.219L367.82 376.22L369.06 376.55C370.01 376.79 370.57 377.76 370.33 378.71C370.09 379.66 369.11 380.22 368.17 379.98L366.91 379.65C361.66 378.26 352.87 375.94 345.77 377.15C339.4 378.24 338.28 381.5 338.28 387.94C338.28 388.92 337.48 389.72 336.5 389.72C336.44 389.72 336.38 389.7 336.32 389.7C336.26 389.71 336.2 389.72 336.14 389.72C335.15 389.72 334.36 388.92 334.36 387.94C334.36 381.49 333.24 378.23 326.88 377.15C319.77 375.94 310.99 378.26 305.74 379.65Z"
+          fill="black"
+        />
+        <path
+          d="M327.31 427.61C319.13 427.51 309.8 428.58 304.23 429.22H304.25C303.26 429.35 302.39 428.63 302.28 427.66C302.17 426.68 302.87 425.8 303.84 425.69C309.49 425.04 318.99 423.96 327.37 424.06C332.3 424.13 334.92 426.03 336.32 428.75C337.72 426.03 340.34 424.13 345.27 424.06C353.68 423.96 363.14 425.04 368.8 425.69C369.77 425.8 370.47 426.69 370.36 427.66C370.25 428.63 369.38 429.35 368.39 429.22L368.331 429.213L368.329 429.213C362.752 428.572 353.51 427.51 345.31 427.61C339.29 427.69 338.27 430.69 338.27 439.67C338.27 440.65 337.47 441.45 336.49 441.45C336.46 441.45 336.43 441.445 336.4 441.44C336.386 441.438 336.372 441.435 336.359 441.434C336.342 441.431 336.326 441.43 336.31 441.43C336.25 441.44 336.19 441.45 336.13 441.45C335.14 441.45 334.35 440.65 334.35 439.67C334.35 430.69 333.34 427.69 327.31 427.61Z"
+          fill="black"
+        />
+        <path
+          d="M336.49 501.6H336.13C335.15 501.6 334.35 502.4 334.35 503.38V570.51C334.35 571.49 335.14 572.29 336.13 572.29H336.49C337.47 572.29 338.27 571.49 338.27 570.51V503.38C338.27 502.4 337.48 501.6 336.49 501.6Z"
+          fill="black"
+        />
+        <path
+          d="M378.16 420.56H378.13L376.66 420.58L376.512 420.582C361.637 420.822 347.627 421.047 340.59 417.06C338.87 416.09 338.25 414.03 338.09 410.93C338.17 409.41 338.14 407.78 338.09 406.08C338.099 405.632 338.116 405.185 338.134 404.715L338.14 404.55L338.156 404.073C338.212 402.468 338.27 400.798 338.27 399.07C338.27 398.09 337.48 397.29 336.49 397.29C336.43 397.29 336.37 397.3 336.31 397.31C336.293 397.31 336.275 397.308 336.258 397.306C336.25 397.305 336.242 397.304 336.235 397.302L336.22 397.3C336.19 397.295 336.16 397.29 336.13 397.29C335.15 397.29 334.35 398.09 334.35 399.07C334.35 400.97 334.41 402.8 334.48 404.55L334.485 404.67C334.503 405.157 334.521 405.627 334.53 406.08C334.48 407.78 334.45 409.41 334.53 410.93C334.36 414.03 333.74 416.09 332.03 417.06C324.97 421.06 310.88 420.83 295.96 420.58L294.49 420.56H294.46C293.49 420.56 292.7 421.34 292.68 422.31C292.66 423.29 293.45 424.1 294.43 424.11L295.9 424.13L296.265 424.136H296.271C299.849 424.194 303.314 424.25 306.62 424.25C318.07 424.25 327.7 423.59 333.78 420.15C334.87 419.53 335.69 418.73 336.31 417.78C336.93 418.73 337.75 419.54 338.84 420.15C344.92 423.59 354.55 424.25 366 424.25C369.327 424.25 372.804 424.194 376.386 424.135L376.72 424.13L378.19 424.11C379.17 424.09 379.95 423.29 379.94 422.31C379.92 421.34 379.13 420.56 378.16 420.56Z"
+          fill="black"
+        />
+        <path
+          d="M363.955 468.911C367.324 468.678 371.163 468.412 376.57 468.22L376.56 468.23C377.56 468.2 378.37 468.96 378.4 469.94C378.44 470.92 377.67 471.74 376.69 471.78C371.336 471.972 367.524 472.238 364.181 472.472L364.18 472.472L363.78 472.5C359.5 472.81 355.9 473.06 350.66 473.06C348.96 473.06 347.09 473.03 344.96 472.97C340.25 472.62 337.69 471.12 336.3 467.84C334.92 471.12 332.37 472.61 327.71 472.96C325.54 473.02 323.64 473.05 321.91 473.05C316.67 473.05 313.08 472.8 308.81 472.5C305.37 472.25 301.47 471.98 295.9 471.78C294.92 471.74 294.15 470.92 294.19 469.94C294.23 468.96 295.08 468.19 296.03 468.23C301.66 468.43 305.59 468.7 309.06 468.95L309.083 468.952C314.642 469.341 319.052 469.65 327.53 469.41C332.91 469.01 334.34 468.23 334.34 451.95C334.34 450.97 335.14 450.17 336.12 450.17C336.18 450.17 336.24 450.19 336.3 450.19L336.326 450.186L336.349 450.182C336.367 450.179 336.386 450.177 336.405 450.174C336.43 450.172 336.455 450.17 336.48 450.17C337.47 450.17 338.26 450.97 338.26 451.95C338.26 468.22 339.69 469 345.15 469.41C353.535 469.639 357.934 469.332 363.49 468.943L363.955 468.911Z"
+          fill="black"
+        />
+        <path
+          d="M368.62 474.46C362.43 474.34 346.75 475.86 341.11 477.14C338.79 477.67 337.29 478.92 336.31 480.57C335.33 478.92 333.83 477.67 331.51 477.14C325.88 475.86 310.18 474.34 304 474.46C303.02 474.48 302.24 475.29 302.26 476.27C302.28 477.25 303.11 478.04 304.07 478.01C309.99 477.88 325.39 479.39 330.72 480.6C333.57 481.25 334.34 484.02 334.53 488.27C334.48 489.54 334.48 490.81 334.48 492.03V492.81C334.48 493.79 335.28 494.59 336.26 494.59H336.38C337.36 494.59 338.16 493.79 338.16 492.81V492.03C338.16 490.82 338.16 489.55 338.11 488.27C338.3 484.02 339.07 481.25 341.92 480.61C347.26 479.4 362.71 477.88 368.57 478.02C369.5 478.04 370.36 477.26 370.38 476.28C370.4 475.3 369.62 474.49 368.64 474.47L368.62 474.46Z"
+          fill="black"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M664.702 571.051C669.891 581.221 670.001 581.437 670.24 586.61L670.16 586.58C670.25 588.64 670.39 591.46 670.97 596.31C672.95 612.74 672.58 618.5 672.34 621.95V622.17C672.12 625.58 667.67 641.16 667.16 642.92C666.96 643.63 666.35 644.13 665.62 644.2C665.41 644.22 660.52 644.62 657.29 641.1C656.15 639.87 655.4 638.32 654.99 636.5C653.37 636.42 650.93 635.97 649.22 634.2C647.59 632.5 647.01 630.1 647.51 627.04C648.29 622.242 648.03 619.772 647.7 616.633L647.7 616.63C647.47 614.49 647.22 612.07 647.29 608.59C647.349 605.725 646.514 604.039 645.467 601.924L645.44 601.87C643.78 598.54 641.74 594.4 642.63 584.75C642.2 584.21 641.76 583.76 641.34 583.41C641.66 589.34 641.68 601.22 637.21 606.28C635.79 607.89 634.04 608.73 632 608.77C622.216 608.94 621.822 599.818 621.364 589.204L621.36 589.12C621.25 586.43 621.13 583.64 620.87 580.74C619.79 572.3 618.49 568.04 616.08 565.09C612.68 560.94 609.78 551.8 615.72 543.63C616.1 542.91 615.27 540.6 613.69 537.62C613.65 537.56 613.62 537.49 613.59 537.43C611.97 534.4 609.61 530.71 606.99 527.27C599.041 518.971 592.612 514.062 586.393 509.312L586.39 509.31L586.365 509.291C577.434 502.468 568.988 496.016 556.65 480.62C543.24 463.87 539.37 455.74 535.28 447.12L535.275 447.109C532.857 442.023 530.357 436.765 525.74 429.33C517.71 417.83 484.85 373.9 480.77 370.11C480.366 369.731 479.921 369.331 479.441 368.899C479.339 368.808 479.235 368.715 479.13 368.62L479.113 368.604C475.312 365.184 469.383 359.848 464.99 351.92C462.82 361.44 456.54 383.35 439.68 415.81C430.39 434.27 433.02 453.79 435.57 472.66L435.624 473.06C436.297 478.045 436.991 483.182 437.41 488.13C438.07 495.9 439.2 506.02 440.52 516.58C440.52 516.59 440.54 516.61 440.54 516.61V516.65C440.56 516.7 440.56 516.76 440.56 516.82C440.57 516.87 440.58 516.93 440.58 516.98V517.02C440.635 517.44 440.688 517.86 440.74 518.28C440.792 518.7 440.845 519.12 440.9 519.54C440.92 519.61 440.93 519.69 440.93 519.77C443.7 541.49 447.17 564.25 449.07 571.75C449.2 572.29 449.4 573.05 449.66 574.02C459.28 610.38 477.6 696.15 464.25 781.5C464.25 781.63 464.24 781.77 464.21 781.9C464.188 781.988 464.169 782.078 464.15 782.17C464.134 782.246 464.118 782.323 464.1 782.4C462 795.6 459.14 808.79 455.37 821.78C455.62 841.779 456.937 846.39 458.096 850.447L458.1 850.46C458.48 851.77 458.87 853.14 459.19 854.92C459.815 858.324 459.928 861.245 460.038 864.065L460.04 864.13C460.16 867.26 460.29 870.5 461.11 874.45C461.31 875.39 461.96 877.23 462.79 879.56L462.791 879.563C468.652 896.046 484.219 939.825 471.4 1012.83C470.719 1020.43 469.762 1028.44 468.836 1036.2L468.76 1036.84C466.31 1057.35 464 1076.72 466.26 1088.92C468.02 1099.42 468.76 1103.98 468.95 1109.6C470.36 1113.85 477.33 1124.93 481.53 1131.59L482.045 1132.41C483.197 1134.23 484.158 1135.76 484.76 1136.77C485.69 1138.34 486.19 1139.5 486.59 1140.43C487.45 1142.44 487.93 1143.55 492.58 1147.31C493.48 1148.04 494.42 1148.78 495.39 1149.54L495.455 1149.59C501.994 1154.75 510.115 1161.16 514.52 1170.91C516.97 1176.81 516.1 1180.15 514.94 1181.92C514.01 1183.34 512.58 1184.27 510.82 1184.64C510.6 1185.58 510.14 1186.77 509.17 1187.72C508.58 1188.3 507.6 1188.97 506.14 1189.19C505.62 1190.43 504.57 1192.27 502.65 1193.39C501.24 1194.21 499.64 1194.5 497.88 1194.23C496.79 1195.41 494.61 1197.32 491.71 1197.65C489.73 1197.89 487.79 1197.36 485.94 1196.09C484.11 1197.53 480.37 1199.93 475.66 1199.93C475.08 1199.93 474.49 1199.9 473.88 1199.82C468.26 1199.1 463.21 1194.99 458.85 1187.62C455.6 1186.47 442.47 1180.93 443.61 1166.95C443.71 1163.79 440.45 1161.77 434.17 1158.2L433.18 1157.64C425.64 1153.34 421.23 1149.53 425.22 1126.47C425.89 1121.12 425.15 1115.29 424.85 1113.34C423.17 1111.82 420.19 1107.09 421.45 1092.98C422.96 1076.06 424.38 1060.08 403.5 993.29C402.52 990 401.17 988.1 399.74 986.1C395.18 979.73 390.47 973.14 396.76 927.39C400.519 893.919 400.24 893.45 396.843 887.735L396.84 887.73L396.819 887.695C396.093 886.471 395.19 884.951 394.17 882.9C393.71 881.98 393.21 881.05 392.69 880.12L392.56 879.9C392.55 879.89 392.55 879.87 392.55 879.87C391.74 878.43 390.88 876.97 389.99 875.46L389.984 875.45C383.855 865.082 376.229 852.181 374.19 825.81C373.29 814.16 371.74 802.45 369.58 790.74C369.11 788.42 368.68 786.2 368.29 784.07C366.11 773.45 363.42 762.84 360.23 752.31C359.142 748.707 357.317 743.68 355.203 737.856L355.19 737.82C348.47 719.31 338.32 691.33 336.91 668.36C336.8 664.7 336.73 661.49 336.67 658.66L336.669 658.601C336.569 653.944 336.479 649.719 336.23 647.06C335.981 649.709 335.891 653.925 335.791 658.572L335.79 658.63C335.73 661.46 335.66 664.68 335.55 668.3C334.131 691.307 324.001 719.253 317.29 737.763L317.258 737.853C315.143 743.679 313.318 748.707 312.23 752.31C309.04 762.83 306.36 773.42 304.18 784.03C303.79 786.19 303.35 788.45 302.87 790.8C300.71 802.49 299.17 814.18 298.27 825.81C296.231 852.181 288.605 865.082 282.476 875.45L282.47 875.46C282.319 875.718 282.168 875.974 282.019 876.229C281.289 877.472 280.584 878.675 279.92 879.87C279.91 879.88 279.91 879.9 279.91 879.9L279.71 880.25C279.21 881.14 278.74 882.02 278.3 882.9C277.28 884.951 276.377 886.471 275.651 887.695L275.63 887.73C272.23 893.44 271.95 893.91 275.72 927.43C282 973.13 277.29 979.72 272.73 986.1L272.703 986.137C271.283 988.123 269.934 990.01 268.98 993.26C248.09 1060.06 249.51 1076.05 251.02 1092.97C252.28 1107.08 249.3 1111.81 247.62 1113.33C247.33 1115.29 246.59 1121.14 247.27 1126.54C251.25 1149.51 246.84 1153.32 239.3 1157.63L238.31 1158.19C232.04 1161.76 228.77 1163.78 228.88 1167.02C230.02 1180.92 216.89 1186.45 213.64 1187.6C209.28 1194.97 204.23 1199.08 198.61 1199.8C198 1199.87 197.41 1199.91 196.83 1199.91C192.11 1199.91 188.37 1197.51 186.55 1196.07C184.7 1197.33 182.75 1197.86 180.78 1197.63C177.88 1197.3 175.7 1195.39 174.61 1194.21C172.85 1194.48 171.26 1194.2 169.84 1193.37C167.92 1192.24 166.87 1190.41 166.35 1189.17C164.89 1188.95 163.92 1188.29 163.32 1187.7C162.35 1186.75 161.89 1185.57 161.67 1184.62C159.91 1184.26 158.48 1183.32 157.55 1181.9C156.4 1180.13 155.53 1176.79 157.95 1170.94C162.382 1161.13 170.521 1154.71 177.076 1149.55L177.11 1149.52C178.08 1148.76 179.02 1148.02 179.92 1147.29C184.554 1143.54 185.037 1142.43 185.901 1140.43L185.91 1140.41L185.948 1140.32C186.341 1139.41 186.829 1138.27 187.74 1136.75C188.44 1135.58 189.59 1133.75 190.97 1131.56C195.17 1124.9 202.15 1113.83 203.55 1109.58C203.74 1103.97 204.48 1099.4 206.24 1088.93C208.5 1076.7 206.19 1057.33 203.74 1036.82L203.636 1035.95C202.718 1028.25 201.774 1020.33 201.08 1012.66C188.29 939.81 203.85 896.02 209.71 879.54L209.784 879.333C210.579 877.101 211.206 875.342 211.39 874.43C212.21 870.48 212.34 867.24 212.46 864.11C212.57 861.27 212.69 858.33 213.31 854.9C213.63 853.12 214.02 851.76 214.4 850.44L214.405 850.421C215.564 846.366 216.88 841.759 217.13 821.76C213.35 808.77 210.5 795.58 208.4 782.38C208.36 782.22 208.33 782.05 208.29 781.88C208.25 781.75 208.25 781.61 208.25 781.48C194.9 696.14 213.23 610.36 222.84 574C223.09 573.03 223.29 572.27 223.43 571.73C225.33 564.23 228.8 541.48 231.57 519.76C231.57 519.67 231.58 519.59 231.6 519.5C231.71 518.66 231.82 517.82 231.92 516.99V516.95C231.92 516.9 231.94 516.84 231.94 516.79C231.95 516.74 231.95 516.68 231.97 516.62V516.58C231.98 516.56 231.98 516.55 231.98 516.54C233.3 505.98 234.43 495.86 235.09 488.1C235.53 483.02 236.24 477.74 236.93 472.63C239.48 453.76 242.11 434.24 232.81 415.76C215.96 383.31 209.68 361.41 207.51 351.89C203.11 359.84 197.17 365.18 193.37 368.59L193.161 368.777C192.648 369.238 192.162 369.675 191.73 370.08C187.65 373.87 154.79 417.8 146.71 429.38C142.144 436.744 139.644 442.002 137.226 447.088L137.22 447.1C133.13 455.72 129.26 463.86 115.85 480.6C103.51 496.01 95.06 502.46 86.11 509.29L86.1075 509.292C79.8882 514.041 73.4589 518.951 65.51 527.25C60.22 534.18 56.01 542.14 56.82 543.67C62.71 551.78 59.82 560.92 56.42 565.07C54.01 568.02 52.71 572.28 51.63 580.79C51.3884 583.534 51.2682 586.232 51.1521 588.838L51.14 589.11L51.1338 589.254C50.6766 599.842 50.2856 608.899 40.5 608.76C38.47 608.72 36.71 607.88 35.29 606.27C30.82 601.21 30.84 589.33 31.16 583.4C30.74 583.76 30.3 584.2 29.87 584.74C30.7594 594.394 28.7126 598.525 27.0631 601.854L27.06 601.86C25.99 604 25.15 605.69 25.21 608.59C25.2782 611.984 25.0327 614.369 24.8164 616.47L24.7997 616.633C24.4698 619.772 24.2102 622.242 24.99 627.04C25.49 630.1 24.91 632.51 23.28 634.2C21.57 635.97 19.14 636.42 17.51 636.5C17.11 638.32 16.35 639.86 15.22 641.1C11.99 644.62 7.08999 644.22 6.88999 644.2C6.15999 644.13 5.54999 643.62 5.34999 642.92C4.83999 641.16 0.379989 625.59 0.149989 622.17V621.95L0.149774 621.947C-0.0702519 618.506 -0.439367 612.735 1.52999 596.31C2.10903 591.55 2.245 588.746 2.34438 586.696L2.34999 586.58C2.59968 581.396 2.69974 581.201 7.93058 570.958L7.94999 570.92L8.12999 570.57C14.9657 557.668 17.9962 548.132 20.9245 538.917L20.93 538.9C23.53 530.68 26 522.93 30.92 514.02C34.5167 506.273 38.7578 495.763 43.2574 484.613L43.42 484.21C54.53 456.67 68.36 422.41 82.08 405.9C90.1787 396.162 95.9981 390.082 100.248 385.642L100.25 385.64C106.06 379.58 108.98 376.53 111.26 371.02C131.486 323.28 140.616 313.684 145.517 308.533L145.52 308.53L145.619 308.424C146.205 307.802 146.694 307.282 147.12 306.78C147.13 306.76 147.15 306.74 147.17 306.72C147.98 305.73 148.53 304.81 148.98 303.36C154.609 281.87 157.826 273.949 159.946 268.728L159.97 268.67L159.99 268.62C160.981 266.194 161.765 264.273 162.47 261.75C168.35 242.55 182.68 218.54 217.6 212.9L217.757 212.874C218.286 212.788 218.671 212.725 218.89 212.67C219.364 212.545 219.978 212.389 220.687 212.21C221.114 212.101 221.575 211.984 222.06 211.86L222.074 211.856C222.911 211.647 223.906 211.398 224.96 211.13C225.04 211.1 225.13 211.08 225.22 211.06C228.72 210.17 232.74 209.12 233.55 208.75C233.959 208.555 235.504 207.348 236.895 206.26L237.01 206.17C240.35 203.55 244.93 199.97 248.34 198.41C250.156 197.586 253.869 196.025 258.279 194.172C258.849 193.932 259.43 193.688 260.02 193.44L260.794 193.115C268.086 190.059 276.104 186.698 279.03 185.23C283.25 183.12 290.69 176.09 292.62 174.23C293.23 166.349 293.511 140.703 293.558 136.413L293.56 136.22C291.46 131.76 290.6 125.47 290.28 122.33C287.51 122.55 285.24 121.95 283.35 120.52C279.42 117.54 277.19 110.85 276.33 99.47C275.74 91.6 277.16 86.25 280.55 83.58C281.74 82.65 283.01 82.18 284.17 81.97C284.06 81.82 283.96 81.64 283.9 81.45L283.889 81.4155C283.579 80.4404 279.154 66.5074 278.05 54.75C277 43.49 280.93 23.29 288.51 17.47C288.66 16.82 288.79 15.6 287.9 14.81C286.7 13.75 286.33 12.26 286.9 10.83C287.94 8.24 291.78 6.56 296.3 6.7C296.65 6.71 296.89 6.68 297.01 6.66C297.021 6.58774 297.032 6.50003 297.043 6.41233C297.052 6.34226 297.061 6.27218 297.07 6.21L297.072 6.19697C297.172 5.36562 297.337 3.99505 298.63 3.05C300.15 1.95 302.69 1.84 307.11 2.7C309.368 2.80802 310.354 2.135 311.402 1.41949L311.46 1.38C312.74 0.490003 314.34 -0.599997 317.22 0.380003C320.83 0.290003 348.59 -0.0999975 365.88 9.96C383.09 19.99 397.69 32.86 387.33 79.46C387.12 80.42 386.17 81.02 385.22 80.81C385.37 81.28 385.52 81.63 385.64 81.89C387.21 81.71 389.78 81.79 392.07 83.59C395.46 86.26 396.88 91.6 396.29 99.48C395.43 110.86 393.2 117.54 389.27 120.53C387.38 121.97 385.11 122.56 382.34 122.34C382.02 125.47 381.16 131.77 379.06 136.23L379.062 136.423C379.109 140.713 379.39 166.359 380 174.24C381.93 176.1 389.36 183.13 393.59 185.24C396.58 186.74 404.894 190.226 412.31 193.334L412.61 193.46C413.505 193.837 414.379 194.205 415.223 194.56C419.237 196.249 422.576 197.653 424.27 198.43C427.68 199.99 432.26 203.57 435.6 206.19C437.03 207.31 438.64 208.57 439.07 208.78C439.87 209.14 443.87 210.19 447.35 211.08H447.37C447.385 211.08 447.398 211.085 447.41 211.09C447.422 211.095 447.435 211.1 447.45 211.1C447.815 211.193 448.175 211.285 448.523 211.374C449.263 211.563 449.955 211.741 450.56 211.89C451.85 212.22 452.97 212.5 453.73 212.7C453.913 212.753 454.212 212.801 454.615 212.865C454.74 212.885 454.875 212.906 455.02 212.93C489.94 218.57 504.27 242.59 510.14 261.74C510.87 264.32 511.65 266.25 512.65 268.7C514.78 273.93 517.99 281.83 523.67 303.47C524.1 304.87 524.64 305.77 525.45 306.74C525.48 306.77 525.5 306.79 525.52 306.82C525.96 307.35 526.48 307.9 527.11 308.56L527.113 308.563C532.014 313.714 541.144 323.31 561.36 371.03C563.632 376.541 566.54 379.58 572.311 385.609L572.37 385.67C576.62 390.11 582.44 396.2 590.53 405.93C604.231 422.407 618.031 456.604 629.133 484.114L629.18 484.23C633.73 495.53 638.04 506.2 641.73 514.16C646.584 522.921 649.044 530.659 651.644 538.839L651.67 538.92C654.6 548.14 657.63 557.67 664.46 570.57L664.65 570.95L664.702 571.051ZM446.725 214.594L446.59 214.56H446.57C428.68 210.99 409.79 215.88 401.26 218.63C405.32 220.18 411.29 222.73 416.55 226.07C427.76 233.19 432.51 239.48 439.09 248.2L439.406 248.617C442.358 252.515 445.701 256.93 450.21 262.23L450.213 262.233C455.184 268.064 465.462 280.122 478.51 288.97C482.01 291.35 487.87 293.58 495.47 295.45C500.19 296 507.21 297.38 514.97 300.85C516.66 301.61 518.32 302.45 519.94 303.34C514.489 282.724 511.4 275.115 509.35 270.065L509.34 270.04C508.3 267.49 507.48 265.47 506.71 262.74C501.15 244.59 487.55 221.79 454.43 216.44L454.241 216.408C453.57 216.295 453.085 216.213 452.81 216.14C452.06 215.94 450.95 215.66 449.67 215.34C448.658 215.082 447.674 214.834 446.725 214.594ZM388.31 397.66C388.36 392.04 388.24 389.56 388.15 387.75L388.13 387.76L388.128 387.714C388.049 386.038 387.991 384.807 388.1 382.23C388.13 381.6 388.16 380.99 388.2 380.39C388.24 379.35 388.31 378.37 388.44 377.43C389.19 369.78 390.8 364.57 392.04 360.59C393.54 355.79 394.36 353.14 392.87 349.77C390.46 344.34 383.07 340.35 375.76 337.34C374.25 336.81 372.75 336.24 371.27 335.59C360.57 331.62 347.06 328.82 341.25 332.87C339.24 334.27 338.26 336.46 338.26 339.55C338.26 340.53 337.46 341.33 336.48 341.33C336.45 341.33 336.42 341.325 336.39 341.32C336.36 341.315 336.33 341.31 336.3 341.31C336.24 341.32 336.18 341.33 336.12 341.33C335.13 341.33 334.34 340.53 334.34 339.55C334.34 336.46 333.36 334.27 331.35 332.87C325.54 328.82 312.02 331.62 301.31 335.59C299.84 336.23 298.35 336.8 296.85 337.33C289.53 340.34 282.14 344.33 279.73 349.77C278.24 353.13 279.06 355.78 280.56 360.59C281.8 364.56 283.41 369.76 284.16 377.38C284.29 378.33 284.36 379.34 284.4 380.39C284.44 380.99 284.47 381.59 284.5 382.22C284.609 384.801 284.551 386.023 284.473 387.694L284.47 387.75C284.38 389.56 284.26 392.04 284.31 397.66C284.41 409.01 286.65 457.95 288.73 490.59C292.07 543.03 324.09 583.05 336.13 583.05C336.19 583.05 336.25 583.07 336.31 583.07C336.37 583.06 336.43 583.05 336.49 583.05C348.53 583.05 380.55 543.03 383.89 490.59C385.97 457.97 388.21 409.03 388.31 397.66ZM324.76 313.85C330.95 306.57 334.36 297.13 334.36 287.26V237.65C334.36 236.92 334.17 220.12 307.28 217.57C291.6 216.08 282.75 218.7 278 220.11C277.67 220.21 277.37 220.3 277.08 220.38C275.83 220.78 265.88 224.04 257.96 229.07C247.301 235.833 242.949 241.598 236.37 250.314L236.35 250.34C233.31 254.38 229.85 258.96 225.11 264.53L225.099 264.542C220.015 270.507 209.541 282.796 196.1 291.91C194.49 293 192.51 294.03 190.17 295.02C205.25 295.46 213.75 304.24 220.55 312.48C221.53 313.67 222.55 314.81 223.6 315.92C223.64 315.95 223.68 315.98 223.71 316.02L224.43 316.8C242.98 335.82 270.84 342.64 295.36 334.1C296.92 333.46 298.48 332.85 300.03 332.28C309.59 328.1 317.91 321.9 324.76 313.85ZM345.16 316.16C341 311.27 338 305.49 336.31 299.26L336.3 299.25C334.61 305.48 331.61 311.26 327.45 316.15C323.53 320.76 319.14 324.79 314.34 328.22C321.29 326.92 328.72 326.7 333.39 329.95C334.62 330.81 335.58 331.86 336.31 333.09C337.04 331.86 338 330.8 339.23 329.95C343.9 326.7 351.33 326.92 358.27 328.22C353.47 324.8 349.08 320.77 345.16 316.16ZM283.1 340.65C280 341.06 276.89 341.29 273.78 341.29L273.76 341.3C262.5 341.3 251.29 338.59 241.13 333.44C246.53 338.38 251.97 343.07 257.09 347.48C266.16 355.29 274.04 362.08 278.94 367.89C278.373 365.574 277.76 363.591 277.225 361.86L277.151 361.621C275.586 356.566 274.454 352.911 276.48 348.33C277.8 345.36 280.15 342.84 283.1 340.65ZM396.12 348.32C398.16 352.91 397.02 356.57 395.44 361.64H395.43L395.365 361.85C394.83 363.581 394.217 365.564 393.65 367.88C398.55 362.06 406.43 355.27 415.5 347.46C420.62 343.05 426.06 338.36 431.46 333.42C421.31 338.58 410.1 341.28 398.83 341.28C395.72 341.28 392.61 341.06 389.51 340.64C392.45 342.83 394.8 345.35 396.12 348.32ZM372.6 332.28C374.15 332.86 375.71 333.46 377.26 334.1C401.77 342.64 429.62 335.82 448.17 316.81C448.302 316.655 448.441 316.507 448.58 316.357C448.684 316.246 448.788 316.135 448.89 316.02C448.906 316.004 448.923 315.989 448.939 315.975C448.963 315.954 448.986 315.934 449.01 315.91C450.06 314.8 451.08 313.66 452.06 312.47C458.86 304.23 467.36 295.46 482.44 295.01C480.1 294.03 478.11 292.99 476.51 291.9C463.089 282.793 452.609 270.504 447.561 264.585L447.54 264.56C442.825 259.027 439.405 254.495 436.385 250.496L436.26 250.33C429.67 241.6 425.31 235.83 414.65 229.06C406.73 224.02 396.8 220.77 395.54 220.37C395.25 220.29 394.95 220.2 394.62 220.1C389.87 218.69 381.03 216.07 365.34 217.56C338.45 220.11 338.26 236.91 338.26 237.62V287.26C338.26 297.13 341.67 306.57 347.86 313.85C354.72 321.9 363.04 328.1 372.6 332.28ZM391.98 188.43C388.26 186.56 382.68 181.64 379.49 178.67V199.35C379.49 209.29 384.79 212.96 395.89 216.73C399.32 215.4 415.82 209.43 434.37 209.74C434.196 209.607 434.016 209.467 433.827 209.32C433.689 209.214 433.547 209.104 433.4 208.99L433.133 208.782C429.95 206.303 425.727 203.012 422.78 201.67C420.78 200.76 416.14 198.81 411.22 196.75C403.67 193.59 395.12 190 391.98 188.43ZM291.01 20.06C284.88 24.11 280.51 42.86 281.59 54.43V54.44C282.38 62.86 285.02 72.71 286.41 77.49C286.51 72.32 286.57 64.95 286.25 59.42C285.8 51.64 289.75 40.06 296.13 35.88C299.23 33.86 302.63 33.69 305.98 35.38C316.52 40.71 345.63 38.38 364.75 30.67C366.86 30.03 368.89 30.39 370.76 31.65C376.86 35.77 381.11 49.44 382.15 60.81C382.87 68.66 383.6 73.68 384.23 76.9C393.32 34.09 380.49 22.62 364.07 13.05C346.86 3.02 317.29 3.95 317 3.96C316.74 3.97 316.52 3.93 316.3 3.84C314.936 3.31224 314.593 3.54799 313.484 4.31027L313.455 4.3303C312.146 5.22943 310.356 6.45924 306.79 6.26C306.71 6.26 306.63 6.25 306.55 6.23C301.57 5.25 300.75 5.91 300.72 5.94C300.676 6.01945 300.64 6.31709 300.611 6.55089C300.607 6.58194 300.604 6.61187 300.6 6.64L300.596 6.67158C300.465 7.80914 300.163 10.4185 296.17 10.26C292.28 10.1 290.3 11.68 290.19 12.19C292.74 14.38 292.32 17.65 291.68 19.24C291.54 19.58 291.31 19.86 291.01 20.06ZM291.55 118.58C292.05 118.48 292.57 118.61 292.98 118.92L292.99 118.93C293.4 119.24 293.65 119.72 293.68 120.23C293.68 120.33 294.27 129.71 296.94 135.06C297.07 135.31 297.13 135.59 297.13 135.87C297.13 135.906 297.129 135.971 297.128 136.06C297.125 136.265 297.12 136.601 297.12 137.04C300.36 141.88 311.91 156.37 336.29 160.74H336.33C360.73 156.37 372.26 141.89 375.5 137.05C375.49 136.41 375.49 136 375.49 135.87C375.49 135.59 375.56 135.31 375.68 135.06C378.35 129.71 378.94 120.32 378.94 120.23C378.97 119.72 379.22 119.24 379.63 118.93C380.03 118.61 380.56 118.49 381.06 118.59C383.74 119.12 385.61 118.85 387.11 117.71C390.14 115.42 391.98 109.37 392.74 99.22C393.35 91.01 391.53 87.72 389.89 86.41C387.91 84.8224 385.46 85.511 385.259 85.5674L385.25 85.57C384.35 85.84 383.41 85.37 383.08 84.51C381.71 82.64 380.15 77.77 378.62 61.12C377.54 49.23 373.21 37.57 368.78 34.58C367.78 33.9 366.88 33.72 365.94 34C345.85 42.11 315.87 44.34 304.39 38.54C302.18 37.42 300.12 37.52 298.09 38.85C292.96 42.2 289.42 52.55 289.8 59.21C290.302 68.1099 289.87 81.6196 289.807 83.5884C289.803 83.7367 289.8 83.8195 289.8 83.83C289.78 84.37 289.51 84.84 289.12 85.15C288.67 85.58 288 85.77 287.36 85.56C287.18 85.5 284.71 84.81 282.72 86.4C281.08 87.71 279.25 91 279.87 99.21C280.64 109.36 282.48 115.41 285.5 117.7C287.01 118.85 288.87 119.12 291.55 118.58ZM296.67 162.51V199.36H296.68C296.68 206.32 294.18 210.78 290.34 213.95C294.96 213.48 300.64 213.36 307.62 214.02C326.54 215.81 333.65 224.52 336.31 230.85C338.98 224.53 346.08 215.82 365.01 214.02C371.99 213.36 377.67 213.48 382.29 213.95C378.45 210.78 375.95 206.32 375.95 199.36V162.49C375.73 154.51 375.6 145.2 375.54 139.96C370.79 146.19 358.9 158.45 336.64 162.44C336.59 162.45 336.54 162.45 336.49 162.45C336.45 162.45 336.41 162.441 336.37 162.432C336.35 162.428 336.33 162.423 336.31 162.42C336.29 162.423 336.27 162.428 336.25 162.432C336.21 162.441 336.17 162.45 336.13 162.45C336.08 162.45 336.03 162.44 335.98 162.44C320.24 159.62 309.84 152.67 303.88 147.34C300.8 144.58 298.58 142 297.08 140.01C297.01 145.26 296.88 154.55 296.67 162.51ZM249.83 201.66C246.8 203.05 242.41 206.48 239.21 208.99L239.2 209C239.053 209.113 238.911 209.223 238.774 209.33C238.585 209.477 238.404 209.617 238.23 209.75C256.77 209.44 273.27 215.41 276.71 216.74C287.81 212.97 293.11 209.3 293.11 199.36V178.67C289.92 181.63 284.34 186.56 280.62 188.42C277.48 189.99 268.94 193.57 261.4 196.73C256.48 198.79 251.83 200.74 249.83 201.66ZM165.89 262.77C165.14 265.46 164.32 267.48 163.28 270.03H163.27C161.21 275.09 158.12 282.7 152.67 303.33C154.29 302.44 155.95 301.6 157.64 300.84C165.4 297.37 172.43 295.99 177.14 295.44C184.74 293.57 190.6 291.33 194.09 288.96C207.13 280.12 217.41 268.06 222.35 262.26C226.994 256.808 230.384 252.317 233.373 248.359L233.532 248.148C240.095 239.46 244.848 233.169 256.04 226.06C261.31 222.71 267.27 220.16 271.33 218.62C262.8 215.87 243.91 210.99 226.02 214.55L225.897 214.581C224.945 214.821 223.956 215.071 222.94 215.33C221.66 215.65 220.55 215.93 219.8 216.13C219.5 216.21 218.95 216.31 218.18 216.43C185.06 221.78 171.46 244.58 165.89 262.77ZM23.9111 600.328L23.92 600.31H23.91L23.9217 600.287C24.8985 598.323 25.9617 596.186 26.38 592.64C24.98 595.6 23.58 597.32 22.38 598.79L22.3685 598.804C20.8944 600.617 19.8281 601.928 19.1 605.03C18.44 609.6 17.74 612.45 17.22 614.53L17.2122 614.562C16.2082 618.637 15.8642 620.033 17.5 629.1C17.74 630.47 17.87 631.75 17.89 632.94C18.89 632.83 20.04 632.51 20.77 631.75C21.58 630.9 21.84 629.52 21.53 627.63C20.67 622.36 20.98 619.41 21.31 616.28C21.54 614.15 21.77 611.94 21.7 608.69C21.6202 604.93 22.7838 602.592 23.9111 600.328ZM48.15 580.45C49.33 571.15 50.84 566.4 53.72 562.87L53.73 562.86C56.33 559.69 58.69 552.23 53.97 545.74C53.02 544.33 53.25 542.15 54.1 539.7C51.82 540.33 49.58 540.64 47.43 540.64C44.46 540.64 41.68 540.04 39.26 538.84C34.66 536.55 31.07 532.02 29.1 526.27C27.2653 530.827 25.8587 535.248 24.4099 539.802L24.3352 540.036C21.4989 548.965 18.2879 559.073 11.32 572.23L11.15 572.56C7.81666 579.087 6.07999 583.827 5.93999 586.78C5.83999 588.9 5.69999 591.79 5.09999 596.77C3.15999 612.87 3.51999 618.44 3.74999 621.76V621.98C3.88999 624.13 6.49999 633.86 8.42999 640.66C9.59999 640.54 11.4 640.11 12.67 638.71C14.31 636.9 14.77 633.88 14.02 629.73C12.25 619.89 12.7 618.07 13.79 613.67C14.32 611.55 14.97 608.91 15.64 604.36C16.58 600.3 18.07 598.47 19.65 596.54C21.37 594.43 23.32 592.03 25.19 585.74C25.2199 585.65 25.2598 585.56 25.2997 585.471C28.0648 579.811 32.5491 578.641 32.8139 578.572L32.82 578.57C33.38 578.43 33.97 578.58 34.4 578.96C34.84 579.34 35.06 579.91 35 580.48C34.48 585.47 34.02 599.44 38.01 603.95C38.77 604.81 39.6 605.22 40.63 605.24H40.84C46.42 605.24 47.14 600.73 47.65 588.99C47.76 586.26 47.88 583.44 48.15 580.45ZM134.138 445.477C136.577 440.351 139.114 435.02 143.8 427.48L143.79 427.46C146 424.29 150.19 418.53 155.17 411.78C147.77 416.9 140.89 418.96 135.04 418.96C131.43 418.96 128.21 418.18 125.52 416.85C117.53 412.91 107.77 401.77 115.4 370.35C115.11 371.02 114.82 371.7 114.53 372.39C112.014 378.477 108.766 381.873 102.907 387.999L102.704 388.211C98.5029 392.601 92.7529 398.611 84.81 408.17C71.4626 424.245 57.7658 458.176 46.7511 485.463L46.72 485.54C42.15 496.88 37.83 507.59 34.1 515.63C33.1 517.45 32.21 519.22 31.39 520.95C32.54 527.7 36.12 533.31 40.82 535.65C44.92 537.69 50.44 537.54 56.07 535.27C58.32 531.07 61.35 526.93 62.79 525.05C62.83 525 62.87 524.95 62.92 524.9C71.0704 516.38 77.6285 511.374 83.9545 506.545L84 506.51C92.77 499.82 101.05 493.49 113.13 478.42C126.286 462.004 130.068 454.044 134.067 445.627L134.07 445.62L134.138 445.477ZM206.64 345.68C206.78 345.34 207.02 345.06 207.31 344.87L207.29 344.86C208.17 342.03 208.87 339.14 209.35 336.23C210.47 329.31 213.45 310.89 203.45 302.75C197.834 298.173 190.031 298.4 181.768 298.639L181.75 298.64C179.35 298.7 174.09 299.09 167.52 300.99C167.508 300.994 167.496 300.998 167.484 301.002C167.439 301.018 167.398 301.032 167.35 301.04C164.77 301.79 161.99 302.77 159.09 304.07C155.79 305.56 152.6 307.33 149.62 309.33C149.15 309.87 148.64 310.41 148.09 310.99L148.07 311.011C145.208 314.023 140.32 319.167 131.79 335.31C105.87 392.72 117.72 409.05 127.1 413.67C135.82 417.97 151.03 415.37 167.58 395.16C167.587 395.146 167.599 395.137 167.609 395.129C167.613 395.126 167.617 395.123 167.62 395.12C177.52 381.99 187.05 369.69 189.37 367.53C189.767 367.164 190.201 366.773 190.662 366.359C190.792 366.241 190.925 366.122 191.06 366L191.121 365.945C195.701 361.815 202.598 355.596 206.64 345.68ZM240.51 473.16C239.79 478.46 239.12 483.47 238.69 488.46V488.47C238.08 495.55 237.01 505.25 235.73 515.62C245.31 517.29 256.06 514.41 262.38 508.35C267.58 503.37 279.14 488.15 274.24 454.7C271.833 438.329 274.993 420.75 277.78 405.243L277.79 405.19C279.53 395.5 281.06 386.97 280.86 380.63C280.8 379.64 280.72 378.69 280.63 377.78C279.73 371.63 268.17 361.67 254.79 350.15C244.18 341.01 232.2 330.67 221.92 319.3C220.49 317.85 219.12 316.32 217.81 314.73C215.85 312.36 213.8 310 211.53 307.85C215.92 317.889 213.8 331.038 212.87 336.808L212.87 336.81C212.29 340.36 211.4 343.9 210.25 347.33C211.26 352.81 216.47 376.53 236.03 414.2C245.839 433.679 243.13 453.747 240.511 473.156L240.51 473.16ZM265.65 992.3C266.78 988.49 268.37 986.26 269.91 984.1L269.88 984.08C274.03 978.27 278.32 972.26 272.23 927.91C269.85 906.74 269.04 898.27 269.67 893.37C262.89 899.91 255.21 903.55 247.7 903.55H247.34C238.79 903.42 231.53 898.71 226.89 890.29C222.44 882.24 219.77 878.89 214.33 877.2C214.02 878.17 213.61 879.34 213.09 880.79L213.085 880.805C207.31 897.067 191.994 940.203 204.64 1012.26C205.38 1020.31 206.36 1028.52 207.31 1036.46L207.326 1036.59C209.8 1057.33 212.135 1076.91 209.78 1089.61C208.03 1100.04 207.31 1104.5 207.14 1110C207.14 1110.16 207.11 1110.31 207.07 1110.46C205.701 1115 198.689 1126.13 194.039 1133.51L194.03 1133.52L193.552 1134.28C192.404 1136.1 191.449 1137.62 190.85 1138.63C190.055 1139.96 189.636 1140.93 189.265 1141.79L189.23 1141.87C188.12 1144.45 187.3 1145.99 182.21 1150.11C181.3 1150.85 180.35 1151.6 179.37 1152.37L179.36 1152.38C172.783 1157.56 165.328 1163.42 161.27 1172.41C159.37 1177.01 160.03 1179.15 160.59 1180.01C161.26 1181.05 162.46 1181.28 163.35 1181.3C164.33 1181.32 165.11 1182.13 165.09 1183.11C165.09 1183.49 165.24 1184.63 165.88 1185.25C166.28 1185.63 166.85 1185.81 167.64 1185.77C168.55 1185.73 169.23 1186.31 169.43 1187.12C169.43 1187.14 170.04 1189.42 171.72 1190.39C172.59 1190.89 173.63 1190.97 174.88 1190.62C175.6 1190.42 176.37 1190.7 176.8 1191.31L176.812 1191.33C176.98 1191.54 178.808 1193.91 181.27 1194.19C182.68 1194.35 184.06 1193.8 185.5 1192.52C186.2 1191.89 187.28 1191.93 187.94 1192.6C187.98 1192.64 192.49 1197.1 198.24 1196.34C202.83 1195.74 207.11 1192.03 210.95 1185.31C211.18 1184.9 211.56 1184.61 212.01 1184.48C212.6 1184.32 226.45 1180.21 225.39 1167.29C225.19 1161.67 230.23 1158.8 236.61 1155.17L237.59 1154.61C243.61 1151.18 247.59 1148.9 243.82 1127.13C242.91 1119.93 244.24 1112.41 244.3 1112.09C244.41 1111.53 244.78 1111.04 245.3 1110.8C245.299 1110.8 245.303 1110.8 245.313 1110.79C245.55 1110.57 248.817 1107.55 247.55 1093.36C246.01 1076.08 244.56 1059.75 265.65 992.3ZM276.77 878.42C277.63 876.88 278.54 875.33 279.49 873.72H279.48C284.94 864.48 291.62 853.17 294.21 831.58C289.22 845.65 280.75 860.4 267.16 860.4H266.84C251.73 860.15 248.76 847.17 246.6 837.7L246.596 837.682C245.657 833.588 244.688 829.359 243.04 828.57C242.44 828.29 241.54 828.44 240.35 829.02C234.67 831.8 229.96 832.18 225.94 830.19C223.89 829.17 222.16 827.6 220.67 825.57C220.3 842.97 219.02 847.48 217.87 851.48L217.87 851.482C217.5 852.771 217.15 853.991 216.86 855.6C216.29 858.79 216.18 861.47 216.07 864.32C215.96 867.21 215.84 870.19 215.21 873.76C222 875.8 225.29 880 230.04 888.59C234.04 895.85 240.22 899.91 247.43 900.02C257.77 900.12 268.56 892.32 276.44 879.01C276.472 878.955 276.511 878.884 276.554 878.806C276.604 878.715 276.661 878.612 276.72 878.51L276.77 878.42ZM300.78 783.42C300.83 783.1 300.89 782.78 300.95 782.47L300.96 782.49C306.519 751.383 303.681 747.941 297.472 740.412L297.47 740.41L297.458 740.396C290.968 732.516 281.165 720.613 272.73 675.04C270.45 662.69 268.81 647.85 267.08 632.14L267.08 632.138C262.21 588.019 256.71 538.19 234.82 522.9C232.15 543.52 228.9 564.96 226.95 572.67L226.859 573.02C226.728 573.528 226.558 574.185 226.35 574.97C216.77 611.2 198.48 696.83 211.96 781.75C212.77 785.48 213.52 789.27 214.24 792.94L214.253 793.006C217.318 808.632 220.21 823.375 227.54 827.01C230.5 828.48 234.19 828.09 238.8 825.83C241.01 824.74 242.96 824.58 244.59 825.36C247.79 826.89 248.8 831.309 250.08 836.908L250.08 836.91L250.08 836.911C252.19 846.161 254.581 856.64 266.91 856.85H267.16C289.56 856.85 296.84 806.51 296.91 806C296.91 805.985 296.915 805.97 296.92 805.955C296.925 805.94 296.93 805.925 296.93 805.91C297.58 801.13 298.35 796.34 299.21 791.56C299.21 791.46 299.22 791.37 299.24 791.27C299.28 791.08 299.317 790.893 299.355 790.705C299.392 790.518 299.43 790.33 299.47 790.14C299.89 787.9 300.32 785.66 300.78 783.42ZM332.09 668.19C332.19 664.65 332.26 661.43 332.32 658.61H332.31L332.315 658.397C332.462 651.52 332.556 647.143 333.1 644.52C318.01 642.44 311.03 623.96 298.7 582.26C285.03 536.01 249.84 522.81 235.9 519.34C259.895 534.897 265.568 586.257 270.587 631.7L270.618 631.981C272.337 647.528 273.962 662.215 276.21 674.37C284.48 719.06 293.94 730.53 300.2 738.12C304.46 743.28 307.29 746.72 307.49 756.16C307.96 754.55 308.43 752.94 308.92 751.33C309.98 747.827 311.695 743.087 313.682 737.594L314.061 736.548C320.701 718.227 330.713 690.604 332.09 668.19ZM336.31 641.2C336.37 641.19 336.43 641.18 336.49 641.18C350.54 641.18 357 626.96 370.51 581.26C382.03 542.28 408.36 525.9 425.63 519.18C418.77 518.12 412.28 515.24 407.78 510.92C402.18 505.55 389.72 489.27 394.87 454.19C397.12 438.9 394.28 422.23 391.61 407.35C390.96 426.43 389.14 464.1 387.44 490.82C384.1 543.24 351.65 586.6 336.49 586.6C336.46 586.6 336.43 586.595 336.4 586.59C336.37 586.585 336.34 586.58 336.31 586.58C336.25 586.59 336.19 586.6 336.13 586.6C320.97 586.6 288.52 543.24 285.18 490.82C283.48 464.08 281.67 426.42 281.01 407.35C278.34 422.23 275.51 438.9 277.75 454.19C282.9 489.27 270.44 505.55 264.84 510.92C260.34 515.24 253.85 518.12 246.99 519.18C264.26 525.89 290.59 542.28 302.11 581.26C315.62 626.95 322.08 641.18 336.13 641.18C336.16 641.18 336.19 641.185 336.22 641.19C336.25 641.195 336.28 641.2 336.31 641.2ZM365.131 756.137C364.674 754.538 364.197 752.929 363.71 751.33C362.62 747.728 360.838 742.81 358.779 737.128L358.61 736.66C351.96 718.34 341.92 690.65 340.54 668.25C340.43 664.66 340.36 661.45 340.3 658.63L340.295 658.412C340.148 651.528 340.054 647.142 339.51 644.52C354.61 642.44 361.58 623.96 373.91 582.26C387.58 536.01 422.77 522.81 436.71 519.35C412.715 534.906 407.042 586.267 402.023 631.71L402.02 631.74L402.004 631.882C400.28 647.479 398.653 662.197 396.4 674.38C388.132 719.067 378.675 730.533 372.415 738.123L372.41 738.13C368.165 743.274 365.337 746.722 365.131 756.137ZM459.218 879.962C458.838 878.885 458.519 877.98 458.27 877.21L458.24 877.2C452.8 878.9 450.13 882.25 445.68 890.3C441.04 898.72 433.78 903.43 425.24 903.56H424.88C417.37 903.56 409.69 899.92 402.91 893.39C403.53 898.29 402.72 906.75 400.35 927.88C394.251 972.262 398.548 978.278 402.698 984.087L402.7 984.09L402.708 984.101C404.245 986.248 405.842 988.476 406.97 992.31C428.056 1059.73 426.601 1076.05 425.061 1093.33L425.06 1093.34C423.73 1108.22 427.38 1110.82 427.42 1110.84C427.95 1111.08 428.21 1111.5 428.31 1112.07L428.319 1112.12C428.444 1112.86 429.671 1120.14 428.8 1127.03C425.02 1148.88 428.99 1151.15 435.01 1154.59L435.99 1155.15C442.37 1158.78 447.41 1161.65 447.22 1167.19C446.15 1180.25 460.45 1184.43 460.59 1184.47C461.04 1184.6 461.42 1184.9 461.65 1185.3C465.49 1192.01 469.77 1195.72 474.36 1196.33C480.12 1197.09 484.62 1192.63 484.67 1192.58C485.33 1191.91 486.41 1191.88 487.11 1192.5C488.56 1193.78 489.92 1194.33 491.34 1194.17C493.91 1193.89 495.79 1191.32 495.81 1191.29C496.25 1190.69 497.01 1190.41 497.73 1190.61C498.98 1190.96 500.02 1190.88 500.89 1190.38C502.57 1189.41 503.18 1187.13 503.18 1187.11C503.39 1186.3 504.06 1185.75 504.97 1185.77C505.76 1185.81 506.34 1185.63 506.73 1185.25C507.38 1184.63 507.52 1183.49 507.52 1183.1C507.52 1182.13 508.29 1181.33 509.26 1181.31C510.14 1181.29 511.34 1181.05 512.02 1180.02C512.59 1179.16 513.25 1177.02 511.32 1172.37C507.28 1163.44 499.83 1157.56 493.25 1152.38C492.968 1152.16 492.686 1151.93 492.406 1151.71C491.725 1151.17 491.055 1150.64 490.41 1150.12C485.32 1146 484.5 1144.46 483.39 1141.88C483.02 1141.01 482.59 1140.01 481.77 1138.64C481.172 1137.63 480.217 1136.11 479.07 1134.29L478.59 1133.53C473.94 1126.16 466.92 1115.01 465.55 1110.47C465.5 1110.32 465.47 1110.17 465.47 1110.01C465.3 1104.51 464.58 1100.05 462.82 1089.59C460.464 1076.9 462.801 1057.32 465.276 1036.59L465.29 1036.47C466.24 1028.53 467.22 1020.32 467.94 1012.41C480.616 940.194 465.29 897.058 459.516 880.806L459.51 880.79C459.409 880.503 459.312 880.227 459.218 879.962ZM456.557 864.251C456.448 861.426 456.345 858.754 455.77 855.6C455.48 853.99 455.13 852.771 454.76 851.481L454.757 851.468C453.617 847.471 452.33 842.952 451.96 825.57C450.47 827.6 448.74 829.17 446.69 830.19C442.67 832.19 437.95 831.8 432.28 829.02C431.09 828.44 430.18 828.29 429.58 828.57C427.935 829.348 426.966 833.575 426.028 837.663L426.014 837.727C423.844 847.195 420.876 860.15 405.78 860.4H405.46C391.87 860.4 383.4 845.67 378.41 831.59C380.99 853.17 387.67 864.48 393.13 873.71L393.2 873.83C394.135 875.416 395.031 876.938 395.88 878.46L395.9 878.5C395.94 878.572 395.979 878.643 396.018 878.713C396.088 878.837 396.156 878.958 396.22 879.08C403.99 892.19 414.7 900.02 424.89 900.02H425.2C432.41 899.91 438.59 895.86 442.59 888.6C447.34 880.01 450.64 875.81 457.42 873.76C456.79 870.19 456.67 867.21 456.56 864.32L456.557 864.251ZM445.69 572.66C443.74 564.94 440.49 543.51 437.82 522.88L437.83 522.87C415.94 538.17 410.44 587.99 405.57 632.12C403.84 647.83 402.2 662.67 399.92 675.02C391.492 720.6 381.674 732.505 375.184 740.375L375.18 740.38L375.179 740.381C368.97 747.91 366.14 751.341 371.7 782.46L371.88 783.45C372.33 785.65 372.75 787.86 373.16 790.06L373.4 791.26C373.42 791.36 373.43 791.47 373.43 791.57C374.29 796.35 375.05 801.13 375.71 805.9C375.71 805.915 375.715 805.93 375.72 805.945C375.725 805.96 375.73 805.975 375.73 805.99C375.8 806.5 383.1 856.84 405.48 856.84H405.73C418.05 856.63 420.45 846.15 422.56 836.9L422.56 836.898C423.84 831.299 424.85 826.88 428.05 825.35C429.68 824.57 431.63 824.73 433.84 825.82C438.45 828.09 442.14 828.47 445.1 827C452.43 823.365 455.322 808.622 458.387 792.996L458.481 792.518C459.175 788.977 459.891 785.331 460.68 781.74C474.16 696.82 455.87 611.2 446.29 574.96C446.03 573.98 445.83 573.21 445.69 572.66ZM432.113 473.107C429.505 453.714 426.809 433.662 436.59 414.22V414.23C456.16 376.54 461.37 352.82 462.38 347.34C461.23 343.9 460.34 340.37 459.76 336.82L459.76 336.818C458.83 331.048 456.71 317.899 461.1 307.86C458.83 310.01 456.78 312.36 454.82 314.74C453.51 316.33 452.14 317.85 450.72 319.31C440.43 330.68 428.45 341.02 417.84 350.16L417.764 350.226C404.416 361.728 392.888 371.662 392 377.79C391.91 378.71 391.83 379.65 391.77 380.64C391.57 386.97 393.1 395.51 394.84 405.19C397.63 420.71 400.79 438.31 398.39 454.69C393.49 488.14 405.05 503.35 410.25 508.34C416.56 514.4 427.31 517.28 436.9 515.61C435.62 505.24 434.55 495.54 433.94 488.46C433.51 483.47 432.84 478.46 432.12 473.16L432.113 473.107ZM505.02 395.13L505.05 395.16C521.6 415.36 536.81 417.97 545.53 413.67C554.91 409.04 566.77 392.71 540.83 335.29C532.29 319.13 527.41 314 524.54 310.99C523.99 310.41 523.48 309.87 523.01 309.33C520.03 307.33 516.84 305.56 513.53 304.07C510.63 302.77 507.85 301.79 505.27 301.04C505.21 301.03 505.16 301.01 505.11 300.99C498.54 299.1 493.28 298.71 490.87 298.64L490.855 298.64C482.601 298.4 474.787 298.173 469.17 302.75C459.173 310.888 462.148 329.3 463.269 336.234L463.27 336.24C463.75 339.14 464.45 342.04 465.33 344.87C465.62 345.07 465.86 345.34 466 345.68C470.06 355.65 476.99 361.88 481.58 366C481.715 366.122 481.849 366.242 481.98 366.36C482.44 366.774 482.873 367.164 483.27 367.53C485.59 369.69 495.13 382 505.02 395.13ZM609.7 524.9C609.75 524.95 609.79 525 609.83 525.05H609.82C611.26 526.94 614.3 531.08 616.55 535.28C622.17 537.54 627.69 537.69 631.79 535.65C636.49 533.31 640.07 527.7 641.22 520.95C640.42 519.25 639.54 517.52 638.56 515.75C634.823 507.683 630.558 497.123 626.036 485.926L625.88 485.54C614.87 458.24 601.16 424.26 587.79 408.18C579.78 398.54 574.01 392.51 569.79 388.11L569.745 388.064C563.854 381.903 560.604 378.505 558.07 372.39C557.78 371.7 557.49 371.02 557.2 370.35C564.84 401.78 555.08 412.92 547.09 416.86C544.39 418.19 541.18 418.97 537.57 418.97C531.72 418.97 524.84 416.91 517.44 411.8C522.4 418.52 526.57 424.26 528.76 427.4C533.54 435.09 536.09 440.44 538.55 445.62L538.554 445.628C542.562 454.055 546.344 462.005 559.49 478.42C571.57 493.49 579.85 499.82 588.62 506.51C594.97 511.35 601.53 516.36 609.7 524.9ZM654.74 632.92C654.76 631.73 654.88 630.45 655.13 629.09H655.11C656.75 620 656.41 618.61 655.39 614.51L655.355 614.367C654.848 612.3 654.154 609.473 653.48 604.86C652.78 601.9 651.71 600.58 650.23 598.77C649.03 597.29 647.64 595.57 646.24 592.62C646.67 596.18 647.73 598.32 648.71 600.29L648.719 600.308C649.846 602.572 651.01 604.91 650.93 608.67C650.87 611.905 651.098 614.11 651.317 616.231L651.325 616.309C651.643 619.421 651.945 622.377 651.1 627.61C650.8 629.5 651.05 630.89 651.86 631.73C652.59 632.49 653.74 632.81 654.74 632.92ZM668.9 621.96V621.74L668.89 621.72C669.11 618.4 669.47 612.83 667.53 596.73C666.94 591.75 666.8 588.85 666.7 586.74C666.56 583.793 664.823 579.057 661.49 572.53L661.31 572.17C654.369 559.074 651.157 548.974 648.325 540.071L648.29 539.96C646.82 535.33 645.39 530.85 643.54 526.24C641.57 531.99 637.98 536.52 633.38 538.81C630.97 540.01 628.18 540.61 625.21 540.61C623.07 540.61 620.83 540.3 618.56 539.68C619.4 542.11 619.64 544.27 618.71 545.66C613.95 552.21 616.31 559.67 618.91 562.84C621.8 566.37 623.3 571.12 624.48 580.36C624.742 583.317 624.863 586.068 624.979 588.721L624.99 588.97C625.5 600.7 626.22 605.22 631.8 605.22H632.01C633.04 605.2 633.87 604.79 634.63 603.93C638.62 599.42 638.16 585.46 637.64 580.46C637.58 579.89 637.81 579.32 638.24 578.94C638.67 578.56 639.26 578.42 639.82 578.55C640.02 578.6 644.55 579.74 647.34 585.45C647.384 585.521 647.413 585.608 647.44 585.69L647.45 585.72C649.308 592.003 651.266 594.405 652.984 596.513L652.99 596.52L653.006 596.54C654.58 598.472 656.063 600.293 656.98 604.19C657.67 608.89 658.32 611.53 658.85 613.65C659.94 618.05 660.39 619.88 658.62 629.71C657.88 633.86 658.33 636.88 659.97 638.69C661.24 640.09 663.04 640.52 664.21 640.64C666.14 633.85 668.75 624.11 668.9 621.96Z"
+          fill="black"
+        />
+      </g>
+    </svg>
+  );
+};
+
+export default FrontBody;
